Refactor : Step2 3차 피드백 반영 - Data 클래스를 struct 타입으로 변경 - 네이밍 수정 - 매개변수 타입 변경(InputDataValidator → Data)
